third_party/blink/web_tests/svg/parser/whitespace-length-invalid-1-expected.txt

CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitycm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitymm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitycm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitymm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitycm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitymm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitycm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitymm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaNpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "NaN%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Infinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-Infinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnordpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "fnord%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitycm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitymm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitycm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitymm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitycm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitymm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitycm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitymm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaNpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " NaN%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Infinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -Infinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnordpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " fnord%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitycm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitymm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityem  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityex  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypx  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityin  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitycm  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitymm  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypt  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypc  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityem\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityex\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypx\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityin\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitycm\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitymm\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypt\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypc\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityem\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityex\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypx\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityin\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitycm\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitymm\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypt\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypc\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity%\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaNpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   NaN%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Infinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -Infinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnordpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   fnord%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityem  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityex  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypx  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityin  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitycm  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitymm  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypt  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypc  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityem …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityex …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypx …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityin …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitycm …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitymm …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypt …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypc …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity%  …".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityem\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityex\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypx\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityin\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitycm\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitymm\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypt\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypc\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity%\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity\r\n\t…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityem\r…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityex\r…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypx\r…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityin\r…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitycm\r…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitymm\r…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypt\r…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypc\r…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity%\r\n…".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaNpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t NaN%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Infinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -Infinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnordpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t fnord%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitycm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitymm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitycm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitymm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitycm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitymm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitycm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitymm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitycm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitymm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitycm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitymm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaNpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fNaN%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fInfinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinityin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitycm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitymm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinitypc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-Infinity%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnordpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\ffnord%\f".
This is a testharness.js-based test.
Harness: the test ran to completion.
